whereisit/requirements.txt
FASTAPI>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
jinja2>=3.1.0
python-multipart>=0.0.9
qrcode[pil]>=7.4.2
Pillow==10.3.0
python-multipart==0.0.9
